Rename ghz_frontend.py to load_test.py. Cleanup args
- Make --experiment and --parallel_experiment optional
(multiplexed is enabled via the config file).
- Add --config option for setting config file name.
- Remove demos.py
